Use floats for represent infinity
Python can convert easily between ints and floats

+++ 2023-Xtreme-Xmas-Code-Maiden-Voyage/Day25-Python-Thorough-Comments/component.py
@@ -8,7 +8,7 @@ class Component:
     def __init__(self, id: str, connected_components: list):
         self.id = id
         self.connected_components = connected_components
-        self.distance = int("inf")
+        self.distance = float("inf")
         self.previous = None
 
     # Return a string representation of the Component object for printing
